more portable typedefs

+++ lib.h
@@ -61,15 +61,15 @@
 #define max(a,b) (((a)>(b))?(a):(b))
 #define clamp(a,x,b) (((x)<(a))?(a):((x)>(b))?(b):(x))
 
-typedef int8_t i8;
-typedef int16_t i16;
-typedef int32_t i32;
-typedef int64_t i64;
+typedef int_least8_t i8;
+typedef int_least16_t i16;
+typedef int_least32_t i32;
+typedef int_least64_t i64;
 
-typedef uint8_t u8;
-typedef uint16_t u16;
-typedef uint32_t u32;
-typedef uint64_t u64;
+typedef uint_least8_t u8;
+typedef uint_least16_t u16;
+typedef uint_least32_t u32;
+typedef uint_least64_t u64;
 
 typedef float f32;
 typedef double f64;
@@ -77,7 +77,7 @@ typedef double f64;
 typedef size_t usize;
 typedef uintptr_t uptr;
 typedef ptrdiff_t dptr;
-typedef i32 bool;
+typedef u8 bool;
 
 typedef struct {
 	const u8 *str;
